fix: use tabpage numbers instead of tab ids
In neovim, there is a disconnect between the tab number and tab id.
What is actually important is the tab number because this defines the
order of tabs displayed. This is especially true when restoring a
session because when neovim restores tabs, the tab ids are very likely
to be different. Neovim will restore order of tabs which means tabpage
numbers will be the same across sessions. Therefore using tabpage number
is the only correct key to use.